+++ instrumentation/google.golang.org/grpc/otelgrpc/test/stats_handler_test.go
@@ -114,7 +114,7 @@ func assertStatsHandlerServerMetrics(t *testing.T, reader metric.Reader, service
 			},
 			{
 				Name:        "rpc.server.responses_per_rpc",
-				Description: "Measures the number of messages responded per RPC. Should be 1 for all non-streaming RPCs.",
+				Description: "Measures the number of messages received per RPC. Should be 1 for all non-streaming RPCs.",
 				Unit:        "{count}",
 				Data: metricdata.Histogram[int64]{
 					Temporality: metricdata.CumulativeTemporality,